# specs/001-ai-image-generator/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI Image Generator API
  description: AI 기반 이미지 생성 및 편집 서비스 API
  version: 1.0.0

servers:
  - url: http://localhost:8000/api/v1
    description: Development server

tags:
  - name: auth
    description: 인증 관련 엔드포인트
  - name: jobs
    description: 이미지 생성 작업 관리
  - name: images
    description: 이미지 조회 및 다운로드
  - name: presets
    description: 프리셋 관리

paths:
  /auth/register:
    post:
      tags: [auth]
      summary: 회원가입
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        201:
          description: 회원가입 성공
        400:
          description: 잘못된 요청

  /auth/login:
    post:
      tags: [auth]
      summary: 로그인
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: 로그인 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  token_type:
                    type: string
        401:
          description: 인증 실패

  /auth/refresh:
    post:
      tags: [auth]
      summary: 토큰 갱신
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        200:
          description: 토큰 갱신 성공
        401:
          description: 리프레시 토큰 만료

  /auth/me:
    get:
      tags: [auth]
      summary: 현재 사용자 정보 조회
      security:
        - bearerAuth: []
      responses:
        200:
          description: 사용자 정보
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  email:
                    type: string
                  daily_usage:
                    type: object
                    properties:
                      used:
                        type: integer
                      limit:
                        type: integer
                      remaining:
                        type: integer

  /jobs:
    post:
      tags: [jobs]
      summary: 이미지 생성 작업 생성
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [type, prompt]
              properties:
                type:
                  type: string
                  enum: [text2img, img2img, inpaint]
                prompt:
                  type: string
                negative_prompt:
                  type: string
                aspect_ratio:
                  type: string
                  enum: ["1:1", "4:3", "3:4", "16:9", "9:16"]
                seed:
                  type: integer
                  default: -1
                steps:
                  type: integer
                  default: 30
                guidance_scale:
                  type: number
                  default: 7.5
                strength:
                  type: number
                  default: 0.75
                input_image:
                  type: string
                  format: binary
                mask_image:
                  type: string
                  format: binary
      responses:
        201:
          description: 작업 생성 성공
        400:
          description: 잘못된 요청
        429:
          description: 일일 생성 횟수 초과

    get:
      tags: [jobs]
      summary: 작업 목록 조회
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: 작업 목록

  /jobs/{jobId}:
    get:
      tags: [jobs]
      summary: 작업 상세 조회
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: 작업 상세 정보
        404:
          description: 작업 없음

  /images:
    get:
      tags: [images]
      summary: 갤러리 조회
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: 이미지 목록

  /images/{imageId}:
    get:
      tags: [images]
      summary: 이미지 상세 조회
      security:
        - bearerAuth: []
      parameters:
        - name: imageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: 이미지 상세

  /images/{imageId}/download:
    get:
      tags: [images]
      summary: 이미지 다운로드
      security:
        - bearerAuth: []
      parameters:
        - name: imageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: 이미지 파일
          content:
            image/png:
              schema:
                type: string
                format: binary

  /presets:
    get:
      tags: [presets]
      summary: 프리셋 목록 조회
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [text2img, img2img, inpaint]
      responses:
        200:
          description: 프리셋 목록

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT